+++ variables.tf
@@ -56,7 +56,7 @@ variable "enable_helm_wandb" {
 variable "database_engine_version" {
   description = "Version for MySQL Aurora"
   type        = string
-  default     = "8.0.mysql_aurora.3.07.1"
+  default     = "8.0"
 }
 
 variable "database_instance_class" {
@@ -388,6 +388,12 @@ variable "kubernetes_max_nodes_per_az" {
   default     = null
 }
 
+variable "kubernetes_node_disk_size_gb" {
+  type        = number
+  description = "Size of the node root volume in GB."
+  default     = null
+}
+
 variable "eks_policy_arns" {
   type        = list(string)
   description = "Additional IAM policy to apply to the EKS cluster"
@@ -563,14 +569,10 @@ variable "use_ctrlplane_redis" {
   default     = false
 }
 
-variable "cache_size" {
-  description = "Size of the redis cache, when use_ctrlplane_redis is true. These values map to preset sizes in the bitnami helm chart."
-  type        = string
-  default     = "nano"
-  validation {
-    condition     = contains(["nano", "micro", "small", "medium", "large", "xlarge", "2xlarge"], var.cache_size)
-    error_message = "Invalid value specified for 'cache_size'; must be one of 'nano', 'micro', 'small', 'medium', 'large'"
-  }
+variable "use_chainguard_redis" {
+  description = "Whether CHAINGUARD redis is deployed in the cluster"
+  type        = bool
+  default     = false
 }
 
 ##########################################
@@ -620,4 +622,40 @@ variable "clickhouse_endpoint_service_id" {
   type        = string
   description = "The service ID of the VPC endpoint service for Clickhouse"
   default     = ""
+}
+
+variable "external_dns_image_repository" {
+  type        = string
+  description = "The image repository of the external-dns to deploy."
+  default     = "registry.k8s.io/external-dns/external-dns"
+}
+
+variable "external_dns_image_tag" {
+  type        = string
+  description = "The tag of the external-dns to deploy."
+  default     = null
+}
+
+variable "aws_loadbalancer_controller_image_repository" {
+  type        = string
+  description = "The image repository of the aws-loadbalancer-controller to deploy."
+  default     = "public.ecr.aws/eks/aws-load-balancer-controller"
+}
+
+variable "aws_loadbalancer_controller_image_tag" {
+  type        = string
+  description = "The tag of the aws-loadbalancer-controller to deploy."
+  default     = null
+}
+
+variable "cluster_autoscaler_image_repository" {
+  type        = string
+  description = "The image repository of the cluster-autoscaler to deploy."
+  default     = "registry.k8s.io/autoscaling/cluster-autoscaler"
+}
+
+variable "cluster_autoscaler_image_tag" {
+  type        = string
+  description = "The tag of the cluster-autoscaler to deploy."
+  default     = null
 }