+++ versions.tf
@@ -9,9 +9,17 @@ terraform {
       source  = "hashicorp/kubernetes"
       version = "~> 2.23"
     }
+    helm = {
+      source  = "hashicorp/helm"
+      version = "< 3.0.0"
+    }
     time = {
       source  = "hashicorp/time"
       version = "~> 0.13"
+    }
+    null = {
+      source  = "hashicorp/null"
+      version = "~> 3.0"
     }
   }
 }